feat/add-user-balance

+++ internal/domain/gophermart/entities/entities.go
@@ -16,7 +16,7 @@ type UserAuth struct {
 type Order struct {
 	Number     string    `json:"number" db:"order_number"`
 	Status     string    `json:"status" db:"status"`
-	Accrual    *int      `json:"accrual,omitempty" db:"accrual"`
+	Accrual    *float64  `json:"accrual,omitempty" db:"accrual"`
 	UploadedAt time.Time `json:"uploaded_at" db:"upload_time"`
 }
